requirements.txt
torch>=2.0
torchvision>=0.15
peft>=0.7
timm>=0.9
numpy>=1.24
matplotlib>=3.7
wandb>=0.16
faiss-cpu>=1.7
scikit-learn>=1.3
pyyaml>=6.0
